pyproject.toml
[tool.poetry]
name = "ProjectorPi"
version = "0.1.0"
description = ""
authors = ["Marijn Doeve <marijn@doeve.me>"]

[tool.poetry.dependencies]
python = "^3.9"
pyserial = "^3.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
